fix autoupdate refresh for goals
